fix: Revert dependent fields refactoring (#1900)
* Revert "fix: Fix issue in edit expense and add refactoring changes to mileage/per-diem (#1861)"

This reverts commit 66c7754a9aaa50cfe65aba41034f8f8befe28a70.

* Revert "feat: Dependent Fields: Create new component for dependent fields (#1785)"

This reverts commit 495e192b304baabb5b538e030f3c6f9d64bae335.
