+++ src/app/fyle/add-edit-expense/add-edit-expense.page.html
@@ -408,24 +408,30 @@
             </div>
           </ng-container>
 
-          <ng-container *ngIf="txnFields$|async as txnFields">
-            <ng-container *ngIf="dependentFields$|async as dependentFields">
-              <ng-container *ngIf="etxn$|async as etxn">
-                <ng-container *ngIf="selectedProject$|async as selectedProject">
-                  <ng-container formArrayName="dependent_fields">
-                    <app-dependent-fields
-                      #dependentFieldsRef
-                      [dependentFieldsFormArray]="fg.controls.dependent_fields"
-                      [dependentCustomFields]="dependentFields"
-                      [parentFieldId]="txnFields?.project_id?.id"
-                      [parentFieldValue]="selectedProject?.projectv2_name"
-                      [txnCustomProperties]="etxn.tx.custom_properties"
-                    ></app-dependent-fields>
-                  </ng-container>
-                </ng-container>
-              </ng-container>
+          <div *ngIf="dependentFields?.length">
+            <ng-container formArrayName="dependent_fields">
+              <div *ngFor="let dependentField of dependentFields" class="add-edit-expense--internal-block">
+                <div [formGroup]="dependentField.control">
+                  <app-dependent-field
+                    [label]="dependentField.field"
+                    [placeholder]="dependentField.placeholder"
+                    [mandatory]="dependentField.mandatory"
+                    [valid]="!(dependentField?.control?.touched && !dependentField?.control?.valid)"
+                    [fieldId]="dependentField.id"
+                    [parentFieldId]="dependentField.parentFieldId"
+                    [parentFieldValue]="dependentField.parentFieldValue"
+                    formControlName="value"
+                  >
+                  </app-dependent-field>
+                </div>
+              </div>
             </ng-container>
-          </ng-container>
+          </div>
+
+          <div class="add-edit-expense--primary-block" *ngIf="isDependentFieldLoading">
+            <ion-skeleton-text class="add-edit-expense__skeleton-loader--title" [animated]="true"></ion-skeleton-text>
+            <ion-skeleton-text class="add-edit-expense__skeleton-loader" [animated]="true"></ion-skeleton-text>
+          </div>
 
           <ng-container *ngIf="txnFields$|async as txnFields">
             <ng-container *ngIf="!((isProjectsEnabled$|async) && !(isConnected$|async))">
