+++ src/app/fyle/add-edit-expense/add-edit-expense.page.ts
@@ -4,9 +4,11 @@ import { Component, ElementRef, EventEmitter, HostListener, OnInit, ViewChild } 
 import {
   combineLatest,
   concat,
+  EMPTY,
   forkJoin,
   from,
   iif,
+  merge,
   Observable,
   of,
   BehaviorSubject,
@@ -20,7 +22,10 @@ import { TitleCasePipe } from '@angular/common';
 import {
   catchError,
   concatMap,
+  debounceTime,
+  delay,
   distinctUntilChanged,
+  distinctUntilKeyChanged,
   filter,
   finalize,
   map,
@@ -55,6 +60,7 @@ import { CustomInputsService } from 'src/app/core/services/custom-inputs.service
 import { CustomFieldsService } from 'src/app/core/services/custom-fields.service';
 import { cloneDeep, isEqual, isNull, isNumber, mergeWith } from 'lodash';
 import { TransactionService } from 'src/app/core/services/transaction.service';
+import { DataTransformService } from 'src/app/core/services/data-transform.service';
 import { PolicyService } from 'src/app/core/services/policy.service';
 import { TransactionsOutboxService } from 'src/app/core/services/transactions-outbox.service';
 import { LoaderService } from 'src/app/core/services/loader.service';
@@ -97,6 +103,7 @@ import { DuplicateSet } from 'src/app/core/models/v2/duplicate-sets.model';
 import { Expense } from 'src/app/core/models/expense.model';
 import { AccountOption } from 'src/app/core/models/account-option.model';
 import { AccountType } from 'src/app/core/enums/account-type.enum';
+import { LaunchDarklyService } from 'src/app/core/services/launch-darkly.service';
 import { ExpenseType } from 'src/app/core/enums/expense-type.enum';
 import { PaymentModesService } from 'src/app/core/services/payment-modes.service';
 import { OrgUserSettingsService } from 'src/app/core/services/org-user-settings.service';
@@ -106,9 +113,10 @@ import { ExpensePolicy } from 'src/app/core/models/platform/platform-expense-pol
 import { FinalExpensePolicyState } from 'src/app/core/models/platform/platform-final-expense-policy-state.model';
 import { PublicPolicyExpense } from 'src/app/core/models/public-policy-expense.model';
 import { BackButtonActionPriority } from 'src/app/core/models/back-button-action-priority.enum';
+import { DependentFieldsService } from 'src/app/core/services/dependent-fields.service';
+import { CustomProperty } from 'src/app/core/models/custom-properties.model';
 import { ExpenseField } from 'src/app/core/models/v1/expense-field.model';
 import { StorageService } from 'src/app/core/services/storage.service';
-import { DependentFieldsComponent } from 'src/app/shared/components/dependent-fields/dependent-fields.component';
 
 @Component({
   selector: 'app-add-edit-expense',
@@ -123,8 +131,6 @@ export class AddEditExpensePage implements OnInit {
   @ViewChild('comments') commentsContainer: ElementRef;
 
   @ViewChild('fileUpload', { static: false }) fileUpload: any;
-
-  @ViewChild('dependentFieldsRef') dependentFieldsRef: DependentFieldsComponent;
 
   etxn$: Observable<any>;
 
@@ -347,11 +353,13 @@ export class AddEditExpensePage implements OnInit {
 
   isNewReportsFlowEnabled = false;
 
+  dependentFields = [];
+
+  isDependentFieldLoading = false;
+
   onPageExit$: Subject<void>;
 
   dependentFields$: Observable<ExpenseField[]>;
-
-  selectedProject$: BehaviorSubject<ExtendedProject>;
 
   private _isExpandedView = false;
 
@@ -367,6 +375,7 @@ export class AddEditExpensePage implements OnInit {
     private customInputsService: CustomInputsService,
     private customFieldsService: CustomFieldsService,
     private transactionService: TransactionService,
+    private dataTransformService: DataTransformService,
     private policyService: PolicyService,
     private transactionOutboxService: TransactionsOutboxService,
     private router: Router,
@@ -395,11 +404,17 @@ export class AddEditExpensePage implements OnInit {
     public platform: Platform,
     private titleCasePipe: TitleCasePipe,
     private handleDuplicates: HandleDuplicatesService,
+    private launchDarklyService: LaunchDarklyService,
     private paymentModesService: PaymentModesService,
     private taxGroupService: TaxGroupService,
     private orgUserSettingsService: OrgUserSettingsService,
+    private dependentFieldsService: DependentFieldsService,
     private storageService: StorageService
   ) {}
+
+  get dependentFieldControls() {
+    return this.fg?.controls.dependent_fields as FormArray;
+  }
 
   get isExpandedView() {
     return this._isExpandedView;
@@ -1399,8 +1414,16 @@ export class AddEditExpensePage implements OnInit {
           taxGroups,
           txnFields,
         }) => {
-          if (project) {
-            this.selectedProject$.next(project);
+          const dependentFields: ExpenseField[] = customExpenseFields.filter(
+            (customInput) => customInput.type === 'DEPENDENT_SELECT'
+          );
+
+          if (dependentFields?.length && project) {
+            const projectField = {
+              id: txnFields.project_id?.id,
+              value: project.projectv2_name,
+            };
+            this.addDependentFieldWithValue(etxn.tx.custom_properties, dependentFields, projectField);
           }
 
           const customInputs = this.customFieldsService.standardizeCustomFields(
@@ -2304,8 +2327,6 @@ export class AddEditExpensePage implements OnInit {
   ionViewWillEnter() {
     this.isNewReportsFlowEnabled = false;
     this.onPageExit$ = new Subject();
-    this.dependentFieldsRef?.ngOnInit();
-    this.selectedProject$ = new BehaviorSubject(null);
     this.hardwareBackButtonAction = this.platform.backButton.subscribeWithPriority(
       BackButtonActionPriority.MEDIUM,
       () => {
@@ -2353,10 +2374,6 @@ export class AddEditExpensePage implements OnInit {
     this.systemCategories = this.categoriesService.getSystemCategories();
     this.breakfastSystemCategories = this.categoriesService.getBreakfastSystemCategories();
     this.autoSubmissionReportName$ = this.reportService.getAutoSubmissionReportName();
-
-    this.fg.controls.project.valueChanges
-      .pipe(takeUntil(this.onPageExit$))
-      .subscribe((project) => this.selectedProject$.next(project));
 
     if (this.activatedRoute.snapshot.params.bankTxn) {
       const bankTxn =
@@ -2695,6 +2712,29 @@ export class AddEditExpensePage implements OnInit {
     this.getPolicyDetails();
     this.getDuplicateExpenses();
     this.isIos = this.platform.is('ios');
+
+    this.fg.controls.project.valueChanges
+      .pipe(
+        takeUntil(this.onPageExit$),
+        tap(() => {
+          this.dependentFieldControls.clear();
+          this.dependentFields = [];
+        }),
+        filter((project) => !!project),
+        switchMap((project) => {
+          this.isDependentFieldLoading = true;
+          return this.txnFields$.pipe(
+            take(1),
+            switchMap((txnFields) => this.getDependentField(txnFields.project_id.id, project.projectv2_name)),
+            finalize(() => (this.isDependentFieldLoading = false))
+          );
+        })
+      )
+      .subscribe((res) => {
+        if (res?.dependentField) {
+          this.addDependentField(res.dependentField, res.parentFieldValue);
+        }
+      });
   }
 
   generateEtxnFromFg(etxn$, standardisedCustomProperties$, isPolicyEtxn = false) {
@@ -4247,9 +4287,131 @@ export class AddEditExpensePage implements OnInit {
 
   ionViewWillLeave() {
     this.hardwareBackButtonAction.unsubscribe();
-    this.dependentFieldsRef?.ngOnDestroy();
     this.onPageExit$.next(null);
     this.onPageExit$.complete();
-    this.selectedProject$.complete();
+  }
+
+  private getDependentField(
+    parentFieldId: number,
+    parentFieldValue: string
+  ): Observable<{ dependentField: ExpenseField; parentFieldValue: string }> {
+    return this.dependentFields$.pipe(
+      switchMap((dependentCustomFields) => {
+        const dependentField = dependentCustomFields.find(
+          (dependentCustomField) => dependentCustomField.parent_field_id === parentFieldId
+        );
+        if (dependentField && parentFieldValue) {
+          return this.dependentFieldsService
+            .getOptionsForDependentField({
+              fieldId: dependentField.id,
+              parentFieldId,
+              parentFieldValue,
+              searchQuery: '',
+            })
+            .pipe(
+              map((dependentFieldOptions) =>
+                dependentFieldOptions?.length > 0 ? { dependentField, parentFieldValue } : null
+              )
+            );
+        }
+        return of(null);
+      })
+    );
+  }
+
+  private addDependentField(dependentField: ExpenseField, parentFieldValue: string, value = null): void {
+    const dependentFieldControl = this.formBuilder.group({
+      id: dependentField.id,
+      label: dependentField.field_name,
+      parent_field_id: dependentField.parent_field_id,
+      value: [value, (dependentField.is_mandatory || null) && Validators.required],
+    });
+
+    dependentFieldControl.valueChanges
+      .pipe(takeUntil(this.onPageExit$), distinctUntilKeyChanged('value'))
+      .subscribe((value) => {
+        this.onDependentFieldChanged(value);
+      });
+
+    this.dependentFields.push({
+      id: dependentField.id,
+      parentFieldId: dependentField.parent_field_id,
+      parentFieldValue,
+      field: dependentField.field_name,
+      mandatory: dependentField.is_mandatory,
+      control: dependentFieldControl,
+      placeholder: dependentField.placeholder,
+    });
+
+    this.dependentFieldControls.push(dependentFieldControl, { emitEvent: false });
+  }
+
+  private removeAllDependentFields(updatedFieldIndex: number) {
+    //Remove all dependent field controls after the changed one
+    for (let i = this.dependentFields.length - 1; i > updatedFieldIndex; i--) {
+      this.dependentFieldControls.removeAt(i);
+    }
+
+    //Removing fields from UI
+    this.dependentFields = this.dependentFields.slice(0, updatedFieldIndex + 1);
+  }
+
+  private onDependentFieldChanged(data: { id: number; label: string; parent_field_id: number; value: string }): void {
+    const updatedFieldIndex = this.dependentFieldControls.value.findIndex((depField) => depField.label === data.label);
+
+    //If this is not the last dependent field then remove all fields after this one and create new field based on this field.
+    if (updatedFieldIndex !== this.dependentFieldControls.length - 1) {
+      this.removeAllDependentFields(updatedFieldIndex);
+    }
+
+    //Create new dependent field based on this field
+    this.isDependentFieldLoading = true;
+    this.getDependentField(data.id, data.value)
+      .pipe(finalize(() => (this.isDependentFieldLoading = false)))
+      .subscribe((res) => {
+        if (res?.dependentField) {
+          this.addDependentField(res.dependentField, res.parentFieldValue);
+        }
+      });
+  }
+
+  //Recursive method to add dependent fields with value
+  private addDependentFieldWithValue(
+    txCustomProperties: CustomProperty<string>[],
+    dependentFields: ExpenseField[],
+    parentField: { id: number; value: string }
+  ) {
+    //Get dependent field for the field whose id is parentFieldId
+    const dependentField = dependentFields.find((dependentField) => dependentField.parent_field_id === parentField.id);
+
+    if (dependentField) {
+      //Get selected value for dependent field
+      const dependentFieldValue = txCustomProperties.find(
+        (customProp) => customProp.name === dependentField.field_name
+      );
+
+      if (dependentFieldValue?.value) {
+        //Add dependent field with selected value
+        this.addDependentField(dependentField, parentField.value, dependentFieldValue?.value);
+
+        //Add field which is dependent on the depenent field (if present)
+        const currentField = {
+          id: dependentField.id,
+          value: dependentFieldValue?.value,
+        };
+        this.addDependentFieldWithValue(txCustomProperties, dependentFields, currentField);
+      } else {
+        //If the dependent field does not have a value, trigger the onChange event for parent field
+        //This will add a new field(if it exists) for the selected value of parent field
+        this.isDependentFieldLoading = true;
+        this.getDependentField(parentField.id, parentField.value)
+          .pipe(finalize(() => (this.isDependentFieldLoading = false)))
+          .subscribe((res) => {
+            if (res?.dependentField) {
+              this.addDependentField(res.dependentField, res.parentFieldValue);
+            }
+          });
+      }
+    }
   }
 }
