+++ src/app/fyle/add-edit-mileage/add-edit-mileage.page.html
@@ -329,24 +329,29 @@
               </div>
             </ng-container>
 
-            <ng-container *ngIf="txnFields$|async as txnFields">
-              <ng-container *ngIf="dependentFields$|async as dependentFields">
-                <ng-container *ngIf="etxn$|async as etxn">
-                  <ng-container *ngIf="selectedProject$|async as selectedProject">
-                    <ng-container formArrayName="dependent_fields">
-                      <app-dependent-fields
-                        #dependentFieldsRef
-                        [dependentFieldsFormArray]="fg.controls.dependent_fields"
-                        [dependentCustomFields]="dependentFields"
-                        [parentFieldId]="txnFields?.project_id?.id"
-                        [parentFieldValue]="selectedProject?.projectv2_name"
-                        [txnCustomProperties]="etxn.tx.custom_properties"
-                      ></app-dependent-fields>
-                    </ng-container>
-                  </ng-container>
-                </ng-container>
+            <div *ngIf="dependentFields?.length">
+              <ng-container formArrayName="dependent_fields">
+                <div *ngFor="let dependentField of dependentFields" class="add-edit-mileage--internal-block">
+                  <div [formGroup]="dependentField.control">
+                    <app-dependent-field
+                      [label]="dependentField.field"
+                      [mandatory]="dependentField.mandatory"
+                      [valid]="!(dependentField?.control?.touched && !dependentField?.control?.valid)"
+                      [fieldId]="dependentField.id"
+                      [parentFieldId]="dependentField.parentFieldId"
+                      [parentFieldValue]="dependentField.parentFieldValue"
+                      formControlName="value"
+                    >
+                    </app-dependent-field>
+                  </div>
+                </div>
               </ng-container>
-            </ng-container>
+            </div>
+
+            <div class="add-edit-mileage--primary-block" *ngIf="isDependentFieldLoading">
+              <ion-skeleton-text class="add-edit-mileage__skeleton-loader--title" [animated]="true"></ion-skeleton-text>
+              <ion-skeleton-text class="add-edit-mileage__skeleton-loader" [animated]="true"></ion-skeleton-text>
+            </div>
 
             <ng-container *ngIf="filteredCategories$|async as subCategories">
               <ng-container *ngIf="subCategories.length > 0">
