+++ src/app/fyle/add-edit-mileage/add-edit-mileage.page.ts
@@ -4,19 +4,7 @@ import { Component, ElementRef, EventEmitter, HostListener, OnInit, ViewChild } 
 import { ActivatedRoute, Router } from '@angular/router';
 import { AbstractControl, FormArray, FormBuilder, FormControl, FormGroup, Validators } from '@angular/forms';
 import { LoaderService } from 'src/app/core/services/loader.service';
-import {
-  BehaviorSubject,
-  combineLatest,
-  concat,
-  forkJoin,
-  from,
-  iif,
-  Observable,
-  of,
-  Subject,
-  Subscription,
-  throwError,
-} from 'rxjs';
+import { combineLatest, concat, forkJoin, from, iif, Observable, of, Subject, Subscription, throwError } from 'rxjs';
 import {
   catchError,
   concatMap,
@@ -29,6 +17,9 @@ import {
   take,
   takeUntil,
   tap,
+  filter,
+  delay,
+  distinctUntilKeyChanged,
 } from 'rxjs/operators';
 import { cloneDeep, intersection, isEmpty, isEqual, isNumber } from 'lodash';
 import * as dayjs from 'dayjs';
@@ -84,8 +75,9 @@ import { FinalExpensePolicyState } from 'src/app/core/models/platform/platform-f
 import { PublicPolicyExpense } from 'src/app/core/models/public-policy-expense.model';
 import { BackButtonActionPriority } from 'src/app/core/models/back-button-action-priority.enum';
 import { ExpenseField } from 'src/app/core/models/v1/expense-field.model';
+import { DependentFieldsService } from 'src/app/core/services/dependent-fields.service';
+import { CustomProperty } from 'src/app/core/models/custom-properties.model';
 import { StorageService } from 'src/app/core/services/storage.service';
-import { DependentFieldsComponent } from 'src/app/shared/components/dependent-fields/dependent-fields.component';
 
 @Component({
   selector: 'app-add-edit-mileage',
@@ -100,8 +92,6 @@ export class AddEditMileagePage implements OnInit {
   @ViewChild('comments') commentsContainer: ElementRef;
 
   @ViewChild(RouteSelectorComponent) routeSelector: RouteSelectorComponent;
-
-  @ViewChild('dependentFields') dependentFieldsRef: DependentFieldsComponent;
 
   mode = 'add';
 
@@ -235,11 +225,13 @@ export class AddEditMileagePage implements OnInit {
 
   isNewReportsFlowEnabled = false;
 
+  dependentFields = [];
+
+  isDependentFieldLoading = false;
+
   onPageExit$: Subject<void>;
 
   dependentFields$: Observable<ExpenseField[]>;
-
-  selectedProject$: BehaviorSubject<ExtendedProject>;
 
   private _isExpandedView = false;
 
@@ -283,8 +275,13 @@ export class AddEditMileagePage implements OnInit {
     private categoriesService: CategoriesService,
     private orgSettingsService: OrgSettingsService,
     private platform: Platform,
+    private dependentFieldsService: DependentFieldsService,
     private storageService: StorageService
   ) {}
+
+  get dependentFieldControls() {
+    return this.fg?.controls.dependent_fields as FormArray;
+  }
 
   get showSaveAndNext() {
     return this.activeIndex !== null && this.reviewList !== null && +this.activeIndex === this.reviewList.length - 1;
@@ -891,8 +888,6 @@ export class AddEditMileagePage implements OnInit {
   ionViewWillEnter() {
     this.isNewReportsFlowEnabled = false;
     this.onPageExit$ = new Subject();
-    this.dependentFieldsRef?.ngOnInit();
-    this.selectedProject$ = new BehaviorSubject(null);
 
     this.hardwareBackButtonAction = this.platform.backButton.subscribeWithPriority(
       BackButtonActionPriority.MEDIUM,
@@ -926,10 +921,6 @@ export class AddEditMileagePage implements OnInit {
     const today = new Date();
     this.maxDate = dayjs(this.dateService.addDaysToDate(today, 1)).format('YYYY-MM-D');
     this.autoSubmissionReportName$ = this.reportService.getAutoSubmissionReportName();
-
-    this.fg.controls.project.valueChanges
-      .pipe(takeUntil(this.onPageExit$))
-      .subscribe((project) => this.selectedProject$.next(project));
 
     this.fg.reset();
     this.title = 'Add Mileage';
@@ -1418,8 +1409,16 @@ export class AddEditMileagePage implements OnInit {
           recentProjects,
           recentCostCenters,
         }) => {
-          if (project) {
-            this.selectedProject$.next(project);
+          const dependentFields: ExpenseField[] = customExpenseFields.filter(
+            (customInput) => customInput.type === 'DEPENDENT_SELECT'
+          );
+
+          if (dependentFields?.length && project) {
+            const projectField = {
+              id: txnFields.project_id?.id,
+              value: project.projectv2_name,
+            };
+            this.addDependentFieldWithValue(etxn.tx.custom_properties, dependentFields, projectField);
           }
 
           const customInputs = this.customFieldsService.standardizeCustomFields(
@@ -1563,6 +1562,29 @@ export class AddEditMileagePage implements OnInit {
           }, 1000);
         }
       );
+
+    this.fg.controls.project.valueChanges
+      .pipe(
+        takeUntil(this.onPageExit$),
+        tap(() => {
+          this.dependentFieldControls.clear();
+          this.dependentFields = [];
+        }),
+        filter((project) => !!project),
+        switchMap((project) => {
+          this.isDependentFieldLoading = true;
+          return this.txnFields$.pipe(
+            take(1),
+            switchMap((txnFields) => this.getDependentField(txnFields.project_id.id, project.projectv2_name)),
+            finalize(() => (this.isDependentFieldLoading = false))
+          );
+        })
+      )
+      .subscribe((res) => {
+        if (res?.dependentField) {
+          this.addDependentField(res.dependentField, res.parentFieldValue);
+        }
+      });
   }
 
   async showClosePopup() {
@@ -2485,8 +2507,131 @@ export class AddEditMileagePage implements OnInit {
 
   ionViewWillLeave() {
     this.hardwareBackButtonAction.unsubscribe();
-    this.dependentFieldsRef.ngOnDestroy();
     this.onPageExit$.next(null);
     this.onPageExit$.complete();
+  }
+
+  private getDependentField(
+    parentFieldId: number,
+    parentFieldValue: string
+  ): Observable<{ dependentField: ExpenseField; parentFieldValue: string }> {
+    return this.dependentFields$.pipe(
+      switchMap((dependentCustomFields) => {
+        const dependentField = dependentCustomFields.find(
+          (dependentCustomField) => dependentCustomField.parent_field_id === parentFieldId
+        );
+        if (dependentField && parentFieldValue) {
+          return this.dependentFieldsService
+            .getOptionsForDependentField({
+              fieldId: dependentField.id,
+              parentFieldId,
+              parentFieldValue,
+              searchQuery: '',
+            })
+            .pipe(
+              map((dependentFieldOptions) =>
+                dependentFieldOptions?.length > 0 ? { dependentField, parentFieldValue } : null
+              )
+            );
+        }
+        return of(null);
+      })
+    );
+  }
+
+  private addDependentField(dependentField: ExpenseField, parentFieldValue: string, value = null): void {
+    const dependentFieldControl = this.fb.group({
+      id: dependentField.id,
+      label: dependentField.field_name,
+      parent_field_id: dependentField.parent_field_id,
+      value: [value, (dependentField.is_mandatory || null) && Validators.required],
+    });
+
+    dependentFieldControl.valueChanges
+      .pipe(takeUntil(this.onPageExit$), distinctUntilKeyChanged('value'))
+      .subscribe((value) => {
+        this.onDependentFieldChanged(value);
+      });
+
+    this.dependentFields.push({
+      id: dependentField.id,
+      parentFieldId: dependentField.parent_field_id,
+      parentFieldValue,
+      field: dependentField.field_name,
+      mandatory: dependentField.is_mandatory,
+      control: dependentFieldControl,
+      placeholder: dependentField.placeholder,
+    });
+
+    this.dependentFieldControls.push(dependentFieldControl, { emitEvent: false });
+  }
+
+  private removeAllDependentFields(updatedFieldIndex: number) {
+    //Remove all dependent field controls after the changed one
+    for (let i = this.dependentFields.length - 1; i > updatedFieldIndex; i--) {
+      this.dependentFieldControls.removeAt(i);
+    }
+
+    //Removing fields from UI
+    this.dependentFields = this.dependentFields.slice(0, updatedFieldIndex + 1);
+  }
+
+  private onDependentFieldChanged(data: { id: number; label: string; parent_field_id: number; value: string }): void {
+    const updatedFieldIndex = this.dependentFieldControls.value.findIndex((depField) => depField.label === data.label);
+
+    //If this is not the last dependent field then remove all fields after this one and create new field based on this field.
+    if (updatedFieldIndex !== this.dependentFieldControls.length - 1) {
+      this.removeAllDependentFields(updatedFieldIndex);
+    }
+
+    //Create new dependent field based on this field
+    this.isDependentFieldLoading = true;
+    this.getDependentField(data.id, data.value)
+      .pipe(finalize(() => (this.isDependentFieldLoading = false)))
+      .subscribe((res) => {
+        if (res?.dependentField) {
+          this.addDependentField(res.dependentField, res.parentFieldValue);
+        }
+      });
+  }
+
+  //Recursive method to add dependent fields with value
+  private addDependentFieldWithValue(
+    txCustomProperties: CustomProperty<string>[],
+    dependentFields: ExpenseField[],
+    parentField: { id: number; value: string }
+  ) {
+    //Get dependent field for the field whose id is parentFieldId
+    const dependentField = dependentFields.find((dependentField) => dependentField.parent_field_id === parentField.id);
+
+    if (dependentField) {
+      //Get selected value for dependent field
+      const dependentFieldValue = txCustomProperties.find(
+        (customProp) => customProp.name === dependentField.field_name
+      );
+
+      if (dependentFieldValue?.value) {
+        //Add dependent field with selected value
+        this.addDependentField(dependentField, parentField.value, dependentFieldValue?.value);
+
+        //Add field which is dependent on the depenent field (if present
+        const currentField = {
+          id: dependentField.id,
+          value: dependentFieldValue?.value,
+        };
+        this.addDependentFieldWithValue(txCustomProperties, dependentFields, currentField);
+      } else {
+        //If the dependent field does not have a value, trigger the onChange event for parent field
+        //This will add a new field(if it exists) for the selected value of parent field
+        this.isDependentFieldLoading = true;
+        this.getDependentField(parentField.id, parentField.value)
+          .pipe(finalize(() => (this.isDependentFieldLoading = false)))
+          .subscribe((res) => {
+            if (res?.dependentField) {
+              this.addDependentField(res.dependentField, res.parentFieldValue);
+            }
+          });
+      }
+    }
   }
 }
