+++ src/app/shared/components/dependent-field/dependent-field-modal/dependent-field-modal.component.scss
@@ -1,4 +1,4 @@
-@import '../../../../../../assets/styles/fy_select';
+@import '../../../../../assets/styles/fy_select';
 
 .dependent-field-modal {
   color: $pure-black;
