+++ src/app/shared/components/dependent-field/dependent-field.component.scss
@@ -1,4 +1,4 @@
-@import '../../../../../theme/colors.scss';
+@import '../../../../theme/colors.scss';
 
 .dependent-field {
   &__label {
